# apidocs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "\u7edf\u4e00\u9519\u8bef\u54cd\u5e94\u683c\u5f0f\uff08Pinned Error Contract\uff09\u3002\n\n\u8be5\u7ed3\u6784\u6700\u521d\u7528\u4e8e v2\uff0c\u4f46\u5728\u201c\u5408\u5e76 v2 \u5230 v1\uff08\u4ec5\u4fdd\u7559 /api/v1\uff09\u201d\u540e\uff0c\n\u670d\u52a1\u7aef\u5bf9\u5916\u7edf\u4e00\u4f7f\u7528\u6b64\u9519\u8bef\u54cd\u5e94\u683c\u5f0f\uff0c\u4fbf\u4e8e\u5ba2\u6237\u7aef\u4e0e Web \u7aef\u505a\u4e00\u81f4\u7684\u9519\u8bef\u89e3\u6790\u3002",
        "properties": {
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "TodoItemRestoreRequest": {
        "properties": {
          "client_updated_at_ms": {
            "default": 0,
            "title": "Client Updated At Ms",
            "type": "integer"
          }
        },
        "title": "TodoItemRestoreRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Flow Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/todo/items": {
      "get": {
        "operationId": "list_todo_items_api_v1_todo_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "list_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "List Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "include_archived_lists",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived Lists",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Todo Items",
        "tags": [
          "todo"
        ]
      },
      "post": {
        "operationId": "upsert_todo_item_api_v1_todo_items_post",
        "parameters": [
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Todo Item",
        "tags": [
          "todo"
        ]
      }
    },
    "/api/v1/todo/items/bulk": {
      "post": {
        "operationId": "bulk_upsert_todo_items_api_v1_todo_items_bulk_post",
        "parameters": [
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/TodoItemUpsertRequest"
                },
                "title": "Payloads",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bulk Upsert Todo Items",
        "tags": [
          "todo"
        ]
      }
    },
...
    "/api/v1/todo/items/{item_id}/restore": {
      "post": {
        "operationId": "restore_todo_item_api_v1_todo_items__item_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Restore Todo Item",
        "tags": [
          "todo"
        ]
      }
    },
...
  }
}